Unifies codebase for private and public backend using runtime route injection
