+++ Backend/internal/handlers/points_private.go
@@ -7,22 +7,6 @@ import (
 	"github.com/2024-CMPU9010-GROUP-3/PROJECT/internal/util"
 )
 
-type PointsHandler struct{}
-
-func (p *PointsHandler) HandleGetByRadius(w http.ResponseWriter, r *http.Request) {
-	data := util.Placeholder("GET ByRadius")
-	w.Header().Set(contentType, applicationJson)
-	err := json.NewEncoder(w).Encode(data)
-	util.CheckResponseError(err, w)
-}
-
-func (p *PointsHandler) HandleGetPointDetails(w http.ResponseWriter, r *http.Request) {
-	data := util.Placeholder("GET PointDetails")
-	w.Header().Set(contentType, applicationJson)
-	w.WriteHeader(http.StatusOK)
-	err := json.NewEncoder(w).Encode(data)
-	util.CheckResponseError(err, w)
-}
 
 func (p *PointsHandler) HandlePost(w http.ResponseWriter, r *http.Request) {
 	data := util.Placeholder("POST points")
