--- Backend/public/go.mod
@@ -1,3 +1,0 @@
-module github.com/2024-CMPU9010-GROUP-3/PROJECT
-
-go 1.23.1
