--- Backend/public/internal/handlers/handlers.go
@@ -1,4 +1,0 @@
-package handlers
-
-const contentType = "Content-Type"
-const applicationJson = "application/json"
